error handling and rearranging setup
