+++ src/basic_block/cqlin.rs
@@ -3,8 +3,8 @@
 
 use super::{BasicBlock, BasicBlockType, Data, DataEnc, SRS};
 use crate::{
-  setup::{CQLinSetup, CQSetup},
-  util::{self, calc_pow},
+  graph::{CQLinSetup, SetupType},
+  util::{self, calc_pow, convert_to_data},
 };
 use ark_bn254::{Bn254, Fr, G1Affine, G1Projective, G2Affine, G2Projective};
 use ark_ec::pairing::Pairing;
@@ -29,8 +29,8 @@ impl BasicBlock for CQLinBasicBlock {
     format!("CQLin-{}", self.weights_name)
   }
 
-  fn weights_name(&self) -> String {
-    self.weights_name.clone()
+  fn weights_name(&self) -> Result<String, String> {
+    Ok(self.weights_name.clone())
   }
 
   fn run(&self, weights: &ArrayD<Fr>, inputs: &Vec<&ArrayD<Fr>>) -> Vec<ArrayD<Fr>> {
@@ -42,15 +42,111 @@ impl BasicBlock for CQLinBasicBlock {
     vec![b.dot(&a).into_dyn()]
   }
 
+  fn setup(&self, srs: &SRS, model: &ArrayD<Fr>) -> SetupType {
+    let weights = convert_to_data(srs, model);
+    let m = weights.len();
+    let n = weights[0].raw.len();
+    let N = srs.X2P.len() - 1;
+    let m_inv = Fr::from(m as u64).inverse().unwrap();
+    let domain_m = GeneralEvaluationDomain::<Fr>::new(m).unwrap();
+    let domain_n = GeneralEvaluationDomain::<Fr>::new(n).unwrap();
+    let domain_2m = GeneralEvaluationDomain::<Fr>::new(2 * m).unwrap();
+
+    let mut L_H_i_x = srs.X1P[..n].to_vec();
+    util::ifft_in_place(domain_n, &mut L_H_i_x);
+    let mut L_V_i_x_n: Vec<_> = (0..m).into_par_iter().map(|i| srs.X1P[n * i]).collect();
+    util::ifft_in_place(domain_m, &mut L_V_i_x_n);
+    let mut L_V_i_x: Vec<G1Projective> = srs.X1P[..m].into_par_iter().map(|x| (*x).into()).collect();
+    util::ifft_in_place::<G1Projective>(domain_m, &mut L_V_i_x);
+    let L_H_i_x = L_H_i_x.iter().map(|x| (*x).into()).collect();
+    let L_V_i_x_n = L_V_i_x_n.iter().map(|x| (*x).into()).collect();
+    let L_V_i_x = L_V_i_x.iter().map(|x| (*x).into()).collect();
+
+    // Calculate G1 U for R and S
+    let mut temp: Vec<Vec<_>> = (0..n).into_par_iter().map(|i| (0..m).map(|j| srs.X1P[i + n * j]).collect()).collect();
+    temp.par_iter_mut().for_each(|x| util::ifft_in_place(domain_m, x));
+    let mut U: Vec<Vec<_>> = (0..m).into_par_iter().map(|i| (0..n).map(|j| temp[j][i]).collect()).collect();
+    U.par_iter_mut().for_each(|x| util::ifft_in_place(domain_n, x));
+
+    // Calculate U for mn-degree check
+    let mut temp: Vec<Vec<_>> = (0..n).into_par_iter().map(|i| (0..m).map(|j| srs.X1P[N - m * n + i + n * j]).collect()).collect();
+    temp.par_iter_mut().for_each(|x| util::ifft_in_place(domain_m, x));
+    let mut U_P_R: Vec<Vec<_>> = (0..m).into_par_iter().map(|i| (0..n).map(|j| temp[j][i]).collect()).collect();
+    U_P_R.par_iter_mut().for_each(|x| util::ifft_in_place(domain_n, x));
+
+    // Calculate G2 U for M
+    let mut temp: Vec<Vec<G2Projective>> = (0..n).into_par_iter().map(|i| (0..m).map(|j| srs.X2P[i + n * j]).collect()).collect();
+    temp.par_iter_mut().for_each(|x| util::ifft_in_place(domain_m, x));
+    let mut U2: Vec<Vec<_>> = (0..m).into_par_iter().map(|i| (0..n).map(|j| temp[j][i]).collect()).collect();
+    U2.par_iter_mut().for_each(|x| util::ifft_in_place(domain_n, x));
+    let mut V = srs.X1P[m * n - n..m * n].to_vec();
+    util::ifft_in_place(domain_n, &mut V);
+    V.par_iter_mut().for_each(|x| *x *= m_inv);
+
+    let mut srs_star: Vec<Vec<_>> = (0..m).into_par_iter().map(|i| srs.X1P[n * i..n * i + n].to_vec()).collect();
+    srs_star.par_iter_mut().for_each(|x| util::ifft_in_place(domain_n, x));
+    srs_star = (0..n).into_par_iter().map(|i| (0..m).map(|j| srs_star[m - 1 - j][i]).collect()).collect();
+    srs_star.par_iter_mut().for_each(|x| x.append(&mut vec![G1Projective::zero(); m]));
+    srs_star.par_iter_mut().for_each(|x| util::fft_in_place(domain_2m, x));
+
+    let S: Vec<Vec<_>> = (0..m)
+      .into_par_iter()
+      .map(|i| (0..n).map(|j| (U[i][j] * domain_m.element(i).inverse().unwrap() - V[j]) * weights[i].raw[j]).collect())
+      .collect();
+    let S: Vec<_> = S.par_iter().map(|x| x.iter().sum::<G1Projective>()).collect();
+    let S = S.iter().map(|x| (*x).into()).collect();
+
+    let R: Vec<Vec<_>> = (0..m).into_par_iter().map(|i| (0..n).map(|j| U[i][j] * weights[i].raw[j]).collect()).collect();
+    let R: Vec<_> = R.par_iter().map(|x| x.iter().sum::<G1Projective>()).collect();
+    let R = R.iter().map(|x| (*x).into()).collect();
+
+    let P_R: Vec<Vec<_>> = (0..m).into_par_iter().map(|i| (0..n).map(|j| U_P_R[i][j] * weights[i].raw[j]).collect()).collect();
+    let P_R: Vec<_> = P_R.par_iter().map(|x| x.iter().sum::<G1Projective>()).collect();
+    let P_R = P_R.iter().map(|x| (*x).into()).collect();
+
+    // Calculate C for Q
+    let mut C: Vec<Vec<_>> = (0..n).into_par_iter().map(|i| (0..m).map(|j| weights[j].raw[i]).collect()).collect();
+    C.par_iter_mut().for_each(|x| domain_m.ifft_in_place(x));
+
+    // Calculate Q. The C above corresponds to the C in the cqlin paper
+    C.par_iter_mut().for_each(|x| x.append(&mut vec![Fr::zero(); m]));
+    C.par_iter_mut().for_each(|x| domain_2m.fft_in_place(x));
+    let temp: Vec<Vec<_>> = (0..2 * m).into_par_iter().map(|i| (0..n).map(|j| srs_star[j][i] * C[j][i]).collect()).collect();
+    let mut temp: Vec<_> = temp.par_iter().map(|x| x.iter().sum::<G1Projective>()).collect();
+    util::ifft_in_place(domain_2m, &mut temp);
+    let mut temp = temp[m..].to_vec();
+    util::fft_in_place(domain_m, &mut temp);
+    let Q: Vec<_> = (0..m).into_par_iter().map(|i| temp[i] * domain_m.element(i) * m_inv).collect();
+    let Q = Q.iter().map(|x| (*x).into()).collect();
+
+    let M_x = (0..m).into_par_iter().map(|i| (0..n).map(|j| U2[i][j] * weights[i].raw[j]).sum::<G2Projective>()).sum::<G2Projective>().into(); //TODO: Change to msm
+
+    SetupType::CQLin(CQLinSetup {
+      weights,
+      R,
+      Q,
+      S,
+      P_R,
+      L_V_i_x_n,
+      L_V_i_x,
+      L_H_i_x,
+      M_x,
+    })
+  }
+
   fn prove(
     &mut self,
     srs: &SRS,
-    setup: &(Option<&CQLinSetup>, Option<&CQSetup>),
+    setup: &SetupType,
     inputs: &Vec<&ArrayD<Data>>,
     outputs: &Vec<&ArrayD<Data>>,
     rng: &mut StdRng,
   ) -> (Vec<G1Projective>, Vec<G2Projective>) {
-    let setup = setup.0.unwrap();
+    let setup = if let SetupType::CQLin(cqlin_setup) = setup {
+      cqlin_setup
+    } else {
+      panic!("CQLin has an incorrect SetupType");
+    };
     let weights = &setup.weights;
     let l = inputs[0].len();
     let m = weights.len();
