+++ src/basic_block/sum.rs
@@ -2,10 +2,7 @@
 #![allow(non_upper_case_globals)]
 
 use super::{BasicBlock, BasicBlockType, Data, DataEnc, SRS};
-use crate::{
-  setup::{CQLinSetup, CQSetup},
-  util,
-};
+use crate::{graph::SetupType, util};
 use ark_bn254::{Bn254, Fr, G1Affine, G1Projective, G2Affine, G2Projective};
 use ark_ec::pairing::Pairing;
 use ark_ff::Field;
@@ -33,7 +30,7 @@ impl BasicBlock for SumBasicBlock {
   fn prove(
     &mut self,
     srs: &SRS,
-    _setup: &(Option<&CQLinSetup>, Option<&CQSetup>),
+    _setup: &SetupType,
     inputs: &Vec<&ArrayD<Data>>,
     outputs: &Vec<&ArrayD<Data>>,
     _rng: &mut StdRng,
